Wait QMediaPlayer buffering before entering playing state
QMediaPlayer needs to fill its internal buffer before it can play the
stream otherwise it can end up in a stalled state.

This commit also simplifies the logic for when the player needs to be
restarted and remove the need of an internal timer to poll the player
state.

Fixes choppy http playback.