Fix enabling of play queue navigation actions 'next' and 'previous'
To enable or disable 'next' action, just rely on MPD's status value
'nextSongId'. Enable 'previous' action when there is more than one item in
the play queue or one item for which its duration is known and this is more
than 5 seconds. That is to also support the feature to go to the start of a
track if already played 5 seconds of it at least (which was introduced in
version 2.4.0). Corresponding implementation of Mpris and ThumbnailToolBar
updated accordingly.