Remove superfluous line break from cover tooltip
The line break appears to have originally separated textual information
in the tooltip from the cover image. All the textual information was
removed in 2d8566d74332491d79d6ecbc71a5c2d73fc914ba, so the image is now
the only thing in the tooltip and the line break was just creating an
empty space above it.